backend/backend/api.py
from rest_framework import viewsets, serializers, permissions
from rest_framework import routers

from .models import ToDoList


class ToDoListSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = ToDoList
        fields = ["title", "created_at"]


class ToDoListViewSet(viewsets.ModelViewSet):
    queryset = ToDoList.objects.all()
    serializer_class = ToDoListSerializer
    permission_classes = [permissions.IsAuthenticated]


router = routers.DefaultRouter()
router.register(r"lists", ToDoListViewSet)
