backend/requirements.txt
djangorestframework==3.12.4
markdown==3.3.4
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
distlib==0.3.1
Django==3.2
filelock==3.0.12
identify==2.2.2
idna==2.10
nodeenv==1.5.0
pre-commit==2.12.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
djangorestframework-simplejwt==4.6.0
psycopg2-binary>=2.8
